+++ mlir/lib/Test/Transforms/TestPass.cpp
@@ -1,0 +1,62 @@
+// Copyright 2023 Xanadu Quantum Technologies Inc.
+
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+
+//     http://www.apache.org/licenses/LICENSE-2.0
+
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#define DEBUG_TYPE "test"
+
+#include <memory>
+#include <vector>
+
+#include "llvm/Support/Debug.h"
+#include "llvm/Support/Errc.h"
+
+#include "mlir/Dialect/Arith/IR/Arith.h"
+#include "mlir/Dialect/Bufferization/IR/Bufferization.h"
+#include "mlir/Dialect/Index/IR/IndexDialect.h"
+#include "mlir/Dialect/Linalg/IR/Linalg.h"
+#include "mlir/Dialect/MemRef/IR/MemRef.h"
+#include "mlir/Dialect/SCF/IR/SCF.h"
+#include "mlir/Dialect/Tensor/IR/Tensor.h"
+#include "mlir/IR/BuiltinOps.h"
+#include "mlir/IR/SymbolTable.h"
+#include "mlir/Pass/Pass.h"
+#include "mlir/Transforms/GreedyPatternRewriteDriver.h"
+
+using namespace llvm;
+using namespace mlir;
+
+namespace catalyst {
+namespace test {
+
+#define GEN_PASS_DEF_TESTPASS
+#include "Test/Transforms/Passes.h.inc"
+
+struct TestPass : impl::TestPassBase<TestPass> {
+    using TestPassBase::TestPassBase;
+
+    void runOnOperation() final
+    {
+        LLVM_DEBUG(dbgs() << "test pass"
+                          << "\n");
+
+        getOperation()->emitError() << "TestPass-induced error \n";
+
+        return signalPassFailure();
+    }
+};
+
+} // namespace test
+
+std::unique_ptr<Pass> createTestPass() { return std::make_unique<test::TestPass>(); }
+
+} // namespace catalyst